Refactor pf_model
Move model layer to pf_model. The project will have 3 models: pf_model, pf_bot and pf_webapp
