+++ run.py
@@ -1,0 +1,5 @@
+from pf_model.create_db import create_db
+from pf_model.populate_db import populate_db
+
+create_db()
+populate_db()
